Return search indices as ptrdiff_t instead of long
long is 32 bits on LLP64 targets, so an index past LONG_MAX would be
truncated there. ptrdiff_t covers any element index of an int array and
still carries -1 for absent.

+++ search.c
@@ -2,17 +2,17 @@
 
 #include <assert.h>
 
-long search_linear(const int *a, size_t n, int key)
+ptrdiff_t search_linear(const int *a, size_t n, int key)
 {
     size_t i;
     assert(a != NULL || n == 0);
     for (i = 0; i < n; i++)
         if (a[i] == key)
-            return (long)i;
+            return (ptrdiff_t)i;
     return -1;
 }
 
-long search_binary(const int *a, size_t n, int key)
+ptrdiff_t search_binary(const int *a, size_t n, int key)
 {
     size_t lo = 0, hi;
     assert(a != NULL || n == 0);
@@ -22,7 +22,7 @@ long search_binary(const int *a, size_t n, int key)
     while (lo <= hi) {
         size_t mid = lo + (hi - lo) / 2;
         if (a[mid] == key)
-            return (long)mid;
+            return (ptrdiff_t)mid;
         if (a[mid] < key)
             lo = mid + 1;
         else if (mid == 0)
@@ -33,7 +33,7 @@ long search_binary(const int *a, size_t n, int key)
     return -1;
 }
 
-long search_linear_rec(const int *a, size_t n, int key)
+ptrdiff_t search_linear_rec(const int *a, size_t n, int key)
 {
     assert(a != NULL || n == 0);
     if (n == 0)
@@ -41,19 +41,19 @@ long search_linear_rec(const int *a, size_t n, int key)
     if (a[0] == key)
         return 0;
     {
-        long deeper = search_linear_rec(a + 1, n - 1, key);
+        ptrdiff_t deeper = search_linear_rec(a + 1, n - 1, key);
         return (deeper < 0) ? -1 : deeper + 1;
     }
 }
 
-static long binary_rec_at(const int *a, size_t lo, size_t hi, int key)
+static ptrdiff_t binary_rec_at(const int *a, size_t lo, size_t hi, int key)
 {
     size_t mid;
     if (lo > hi)
         return -1;
     mid = lo + (hi - lo) / 2;
     if (a[mid] == key)
-        return (long)mid;
+        return (ptrdiff_t)mid;
     if (a[mid] < key)
         return binary_rec_at(a, mid + 1, hi, key);
     if (mid == 0)
@@ -61,7 +61,7 @@ static long binary_rec_at(const int *a, size_t lo, size_t hi, int key)
     return binary_rec_at(a, lo, mid - 1, key);
 }
 
-long search_binary_rec(const int *a, size_t n, int key)
+ptrdiff_t search_binary_rec(const int *a, size_t n, int key)
 {
     assert(a != NULL || n == 0);
     if (n == 0)
